Add wave-f32 feature: compile a device-compatible f32 blob on an f64 host
postcard is not self-describing, so a serialized CompiledPedal only loads on a
device whose `Wave` width matches the host that built it. The compiler runs on
an f64 host; embedded devices run f32 — so a baked blob failed to deserialize
on-device (postcard SerdeDeCustom), and the firmware never started.

Add a `wave-f32` cargo feature (pedalkernel-rt + pedalkernel) that forces
`Wave = f32` regardless of target. A build.rs can then compile a blob whose
scalar width matches an f32 device. The compiler keeps computing in f64 for
precision; values are cast to `Wave` only at the storage boundary where they
enter the runtime types — so this is "compute in f64, emit f32", not an f32
migration of the compiler math.

  - pedalkernel-rt: `wave-f32` feature; `Wave` cfg also honors it.
  - pedalkernel: `wave-f32` forwards to rt; re-export `pedalkernel_rt::Wave`
    at the crate root so the compiler can name `crate::Wave` at cast sites.
  - ~500 `as crate::Wave` boundary casts across the compiler (spqr, rigid/*,
    blockwise, k_method, bias, passives, model_lookup, lowering passes, …),
    plus a few reverse `as f64` casts where a compiler-side f64 field/return
    receives a now-f32 runtime value.

Behavior-preserving at the default f64 (the casts are identity there; the
workspace `cargo check` is unaffected). Verified end-to-end on the host: with
`wave-f32`, a compiled blob round-trips through `postcard::from_bytes` and
processes finite, non-silent audio — the same reconstruct-and-run the device
does. The f32 blob is also ~half the size (e.g. a RAT K-table blob 68 -> 34 KB).

Note: at the default f64, clippy may flag `unnecessary_cast` at these sites
(the casts are only load-bearing under `wave-f32`).

+++ pedalkernel/src/models.rs
@@ -440,7 +440,7 @@ impl SpiceOpAmpModel {
             output_impedance: 75.0,
             output_capacitance: 20e-12,
             ota_iabc: 100e-6,
-            ota_vt: pedalkernel_rt::SPICE_VT_27C,
+            ota_vt: pedalkernel_rt::SPICE_VT_27C as f64,
             ota_gm: 0.0,
             ota_r_load: 10_000.0,
         }
